requirements.txt
aiortc==1.5.0
aiohttp==3.8.5
python-socketio==5.9.0
python-engineio==4.7.1
RPi.GPIO==0.7.0
av==10.0.0
opencv-python-headless==4.8.1.78
